Changes includes, CMakeLists, functioonality of analyze_result and the abm example.
