+++ cpp/models/secir/CMakeLists.txt
@@ -1,8 +1,6 @@
 add_library(secir
     age_group.h
     infection_state.h
-    analyze_result.h
-    analyze_result.cpp
     implicit_euler.h
     implicit_euler.cpp
     parameter_space.h
